+++ frontend/src/components/QRModal.vue
@@ -3,7 +3,7 @@
         <div class="modal-dialog modal-dialog-centered modal-fullscreen-md-down">
             <div class="modal-content">
                 <div class="modal-header">
-                    <h1 class="modal-title fs-5 text-primary" id="exampleModalToggleLabel">{{ value.type[1] }}</h1>
+                    <h1 class="modal-title fs-5 text-primary">{{ value.type[1] }}</h1>
                     <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                 </div>
                 <div class="modal-body text-center">